+++ numerapi/__init__.py
@@ -4,3 +4,6 @@ try:
     __version__ = pkg_resources.get_distribution(__name__).version
 except:
     __version__ = 'unknown'
+
+
+from numerapi.numerapi import NumerAPI